FIX: Removed commented code and refactored files.
