+++ .gitignore
@@ -12,6 +12,7 @@
 /build
 
 # misc
+.eslintcache
 .DS_Store
 .env
 .env.local
