+++ src/components/header/NavLinkItems.jsx
@@ -2,10 +2,7 @@ import React from "react";
 const NavLinkItems = (props) => {
   return (
     <div className="single-action-btn">
-      <a href="./components/cart/cart.html">
-        {props.svg}
-        {/* <button>{props.badge}</button> */}
-      </a>
+      <a href="./components/cart/cart.html">{props.svg}</a>
     </div>
   );
 };
