+++ src/components/index.js
@@ -3,3 +3,5 @@ export { NavLinkUser } from "./header/NavLinkUser";
 export { FeaturedItems } from "./homepage/FeaturedItems";
 export { Homepage } from "./homepage/Homepage";
 export { Products } from "./products/Products";
+export { useWishlist } from "../Context/Wishlist/Wishlist-Context";
+export { useCart } from "../Context/Cart/Cart-Context";
